+++ hutool-db/src/main/java/cn/hutool/db/dialect/DialectFactory.java
@@ -49,6 +49,8 @@ public class DialectFactory {
 	public final static String DRIVER_DERBY = "org.apache.derby.jdbc.ClientDriver";
 	/** JDBC 驱动 Derby嵌入式 */
 	public final static String DRIVER_DERBY_EMBEDDED = "org.apache.derby.jdbc.EmbeddedDriver";
+	/** JDBC 驱动 HSQLDB */
+	public final static String DRIVER_HSQLDB = "org.hsqldb.jdbc.JDBCDriver";
 	
 	private static Map<DataSource, Dialect> dialectPool = new ConcurrentHashMap<>();
 	private static Object lock = new Object();
@@ -130,6 +132,9 @@ public class DialectFactory {
 		} else if (nameContainsProductInfo.contains("derby")) {
 			// 嵌入式Derby数据库
 			driver = DRIVER_DERBY_EMBEDDED;
+		} else if (nameContainsProductInfo.contains("hsqldb")) {
+			// HSQLDB
+			driver = DRIVER_HSQLDB;
 		}
 
 		return driver;