+++ hutool-db/src/main/java/cn/hutool/db/ds/hikari/HikariDSFactory.java
@@ -2,7 +2,6 @@ package cn.hutool.db.ds.hikari;
 
 import java.util.Collection;
 import java.util.Map;
-import java.util.Properties;
 import java.util.concurrent.ConcurrentHashMap;
 
 import javax.sql.DataSource;
@@ -14,8 +13,10 @@ import cn.hutool.core.collection.CollectionUtil;
 import cn.hutool.core.io.IoUtil;
 import cn.hutool.core.util.StrUtil;
 import cn.hutool.db.DbRuntimeException;
+import cn.hutool.db.dialect.DriverUtil;
 import cn.hutool.db.ds.DSFactory;
 import cn.hutool.setting.Setting;
+import cn.hutool.setting.dialect.Props;
 
 /**
  * HikariCP数据源工厂类
@@ -92,25 +93,23 @@ public class HikariDSFactory extends DSFactory {
 			group = StrUtil.EMPTY;
 		}
 
-		final Properties config = setting.getProperties(group);
+		final Props config = setting.getProps(group);
 		if (CollectionUtil.isEmpty(config)) {
 			throw new DbRuntimeException("No HikariCP config for group: [{}]", group);
 		}
 		
 		// 规范化属性名
-		if (false == config.containsKey("jdbcUrl") && config.containsKey("url")) {
-			config.put("jdbcUrl", config.remove("url"));
+		final String url = config.getAndRemoveStr(KEY_ALIAS_URL);
+		if (StrUtil.isBlank(url)) {
+			throw new DbRuntimeException("No JDBC URL for group: [{}]", group);
 		}
-		if (false == config.containsKey("username") && config.containsKey("user")) {
-			config.put("username", config.remove("user"));
-		}
-		if (false == config.containsKey("password") && config.containsKey("pass")) {
-			config.put("password", config.remove("pass"));
-		}
-		if (false == config.containsKey("driverClassName") && config.containsKey("driver")) {
-			config.put("driverClassName", config.remove("driver"));
-		}
-
+		config.put("jdbcUrl", url);
+		//自动识别Driver
+		final String driver = config.getAndRemoveStr(KEY_ALIAS_DRIVER);
+		config.put("driverClassName", StrUtil.isNotBlank(driver) ? driver : DriverUtil.identifyDriver(url));
+		config.put("username", config.getAndRemoveStr(KEY_ALIAS_USER));
+		config.put("password", config.getAndRemoveStr(KEY_ALIAS_PASSWORD));
+		
 		final HikariDataSource ds = new HikariDataSource(new HikariConfig(config));
 		return ds;
 	}