+++ hutool-core/src/test/java/cn/hutool/core/util/ClassUtilTest.java
@@ -6,8 +6,6 @@ import java.lang.reflect.Method;
 import org.junit.Assert;
 import org.junit.Test;
 
-import cn.hutool.core.lang.Console;
-
 /**
  * {@link ClassUtil} 单元测试
  * 
@@ -15,12 +13,6 @@ import cn.hutool.core.lang.Console;
  *
  */
 public class ClassUtilTest {
-	
-	@Test
-	public void a() {
-		String path = "C:/Users/maple/Desktop/tone.txt";
-		Console.log(path.matches("^[a-zA-z]:/.*"));
-	}
 	
 	@Test
 	public void getClassNameTest() {
@@ -46,7 +38,7 @@ public class ClassUtilTest {
 	@SuppressWarnings("unused")
 	class TestSubClass extends TestClass {
 		private String subField;
-
+		
 		private void privateSubMethod() {
 		}