+++ CHANGELOG.md
@@ -3,6 +3,21 @@
 
 -------------------------------------------------------------------------------------------------------------
 
+## 4.1.21
+
+### 新特性
+* 【core】        RuntimeUtil增加getErrorResult方法（issue#199@Github）
+* 【core】        ReflectUtil增加hasField方法（感谢@【杭州】J辉）
+* 【core】        BeanUtil增加toBean方法（感谢@【杭州】J辉）
+* 【db】           增加对HSQLDB支持，改进Driver自定识别
+
+### Bug修复
+* 【core】        修复EnumUtil.getFieldNames定义name属性重复问题（感谢@【杭州】J辉）
+* 【json】         修复List多层嵌套toBean转换失败问题
+* 【core】        修复ObjectUtil.toString问题（issue#IONLA@Gitee）
+
+-------------------------------------------------------------------------------------------------------------
+
 ## 4.1.20
 
 ### 新特性